+++ src/components/Input/index.tsx
@@ -4,15 +4,22 @@ import "./styles.css";
 interface PropTypes
   extends InputHTMLAttributes<HTMLInputElement | HTMLTextAreaElement> {
   label: string;
+  error: boolean;
+  errors: string[];
+  forceShowError: boolean;
 }
 
 const InputField = ({
   label,
   type = "text",
   placeholder,
+  error,
+  errors,
+  forceShowError,
   ...props
 }: PropTypes) => {
   const [isFocused, setIsFocused] = useState(false);
+  const [touched, setTouched] = useState(false);
 
   return (
     <div className="input-container">
@@ -22,9 +29,16 @@ const InputField = ({
         type={type}
         className={`input ${isFocused ? "focused" : ""}`}
         onFocus={() => setIsFocused(true)}
-        onBlur={() => setIsFocused(false)}
+        onBlur={() => {
+          setIsFocused(false);
+          setTouched(true);
+        }}
         placeholder={placeholder}
+        required={false}
       />
+      {(touched || forceShowError) && error && (
+        <p className="error">{errors[0]}</p>
+      )}
     </div>
   );
 };