+++ src/types/schema.ts
@@ -1,12 +1,12 @@
 import { InputHTMLAttributes } from "react";
 
 export interface FieldValidations {
-  isString: boolean;
-  isNumber: boolean;
-  isEmail: boolean;
-  isPassword: boolean;
-  min: number;
-  max: number;
+  isString?: boolean;
+  isNumber?: boolean;
+  isEmail?: boolean;
+  min?: number;
+  max?: number;
+  required?: boolean;
 }
 
 export interface Field {
@@ -14,11 +14,15 @@ export interface Field {
   type: "input" | "textarea" | "dropdown" | "checkbox" | "radio";
   initialValue?: string | number | boolean;
   label: string;
-  required?: boolean;
   validations?: FieldValidations;
   inputProps?: InputHTMLAttributes<HTMLInputElement | HTMLTextAreaElement>;
 }
 
 export interface Schema {
   fields: Field[];
+}
+
+export interface Error {
+  name: string;
+  message: string;
 }